requirements.txt
torch
openai-whisper
setuptools-rust 
pyaudio
sounddevice
numpy
scipy 
webrtcvad